Move scalar-related operators to their own package
