+++ pkg/streamingpromql/operators/scalars/instant_vector_to_scalar.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package operators
+package scalars
 
 import (
 	"context"
