+++ pkg/streamingpromql/operators/scalars/scalar_constant.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package operators
+package scalars
 
 import (
 	"context"
